requirements.txt
numpy
pandas
pillow
matplotlib
opencv-python
torchvision
mlflow
seaborn
tqdm
lightning
kagglehub
